Refactor digest block creation and ordering
